requirements_brax.txt
absl-py==2.1.0
blinker==1.7.0
brax==0.10.3
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.8
etils==1.7.0
Flask==3.0.2
Flask-Cors==4.0.0
flax==0.8.2
fonttools==4.50.0
fsspec==2024.3.1
glfw==2.7.0
grpcio==1.62.1
gym==0.26.2
gym-notices==0.0.8
imageio==2.34.0
importlib_resources==6.4.0
itsdangerous==2.1.2
jax==0.4.25
jaxlib==0.4.25+cuda11.cudnn86
jaxopt==0.8.3
Jinja2==3.1.3
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.2
msgpack==1.0.8
mujoco==3.1.3
mujoco-mjx==3.1.3
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.7
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
Pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytinyrenderer==0.0.14
PyYAML==6.0.1
rich==13.7.1
scipy==1.12.0
six==1.16.0
tensorboardX==2.6.2.2
tensorstore==0.1.56
toolz==0.12.1
trimesh==4.2.4
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.18.1
